+++ analyzer.go
@@ -114,6 +114,24 @@ func (analyzer *Analyzer) ActiveVisitors(filter *Filter, duration time.Duration)
 	return stats, count, nil
 }
 
+// TotalVisitors returns the total visitor count, session count, bounce rate, and views.
+func (analyzer *Analyzer) TotalVisitors(filter *Filter) (*TotalVisitorStats, error) {
+	args, query := buildQuery(analyzer.getFilter(filter), []field{
+		fieldVisitors,
+		fieldSessions,
+		fieldViews,
+		fieldBounces,
+		fieldBounceRate,
+	}, nil, nil)
+	stats := new(TotalVisitorStats)
+
+	if err := analyzer.store.Get(stats, query, args...); err != nil && err != sql.ErrNoRows {
+		return nil, err
+	}
+
+	return stats, nil
+}
+
 // Visitors returns the visitor count, session count, bounce rate, and views grouped by day.
 func (analyzer *Analyzer) Visitors(filter *Filter) ([]VisitorStats, error) {
 	args, query := buildQuery(analyzer.getFilter(filter), []field{
@@ -451,7 +469,7 @@ func (analyzer *Analyzer) PageConversions(filter *Filter) (*PageConversionsStats
 	})
 	stats := new(PageConversionsStats)
 
-	if err := analyzer.store.Get(stats, query, args...); err != nil {
+	if err := analyzer.store.Get(stats, query, args...); err != nil && err != sql.ErrNoRows {
 		return nil, err
 	}
 
@@ -634,7 +652,7 @@ func (analyzer *Analyzer) Platform(filter *Filter) (*PlatformStats, error) {
 
 	stats := new(PlatformStats)
 
-	if err := analyzer.store.Get(stats, query.String(), args...); err != nil {
+	if err := analyzer.store.Get(stats, query.String(), args...); err != nil && err != sql.ErrNoRows {
 		return nil, err
 	}
 
@@ -987,7 +1005,7 @@ func (analyzer *Analyzer) totalSessionDuration(filter *Filter) (int, error) {
 		)`, filterQuery))
 	var averageTimeSpentSeconds int
 
-	if err := analyzer.store.Get(&averageTimeSpentSeconds, query.String(), args...); err != nil {
+	if err := analyzer.store.Get(&averageTimeSpentSeconds, query.String(), args...); err != nil && err != sql.ErrNoRows {
 		return 0, err
 	}
 
@@ -999,7 +1017,7 @@ func (analyzer *Analyzer) totalEventDuration(filter *Filter) (int, error) {
 	query := fmt.Sprintf(`SELECT sum(duration_seconds) FROM event WHERE %s`, filterQuery)
 	var averageTimeSpentSeconds int
 
-	if err := analyzer.store.Get(&averageTimeSpentSeconds, query, filterArgs...); err != nil {
+	if err := analyzer.store.Get(&averageTimeSpentSeconds, query, filterArgs...); err != nil && err != sql.ErrNoRows {
 		return 0, err
 	}
 
@@ -1057,7 +1075,7 @@ func (analyzer *Analyzer) totalTimeOnPage(filter *Filter) (int, error) {
 		AverageTimeSpentSeconds int `db:"average_time_spent_seconds" json:"average_time_spent_seconds"`
 	})
 
-	if err := analyzer.store.Get(stats, query.String(), args...); err != nil {
+	if err := analyzer.store.Get(stats, query.String(), args...); err != nil && err != sql.ErrNoRows {
 		return 0, err
 	}