+++ model.go
@@ -133,6 +133,15 @@ type ActiveVisitorStats struct {
 	Visitors int    `json:"visitors"`
 }
 
+// TotalVisitorStats is the result type for total visitor statistics.
+type TotalVisitorStats struct {
+	Visitors   int     `json:"visitors"`
+	Views      int     `json:"views"`
+	Sessions   int     `json:"sessions"`
+	Bounces    int     `json:"bounces"`
+	BounceRate float64 `db:"bounce_rate" json:"bounce_rate"`
+}
+
 // VisitorStats is the result type for visitor statistics.
 type VisitorStats struct {
 	Day        time.Time `json:"day"`